requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
astroid==3.2.4
attrs==24.2.0
bleach==6.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.1
cryptography==43.0.3
dill==0.3.8
feedparser==6.0.11
Flask==3.0.3
Flask-Mail==0.10.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
http_ece==1.2.1
idna==3.8
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.1
platformdirs==4.2.2
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
py-vapid==1.9.1
pycparser==2.22
pylint==3.2.7
PySocks==1.7.1
python-dotenv==1.0.1
pywebpush==2.0.1
requests==2.32.3
selenium==4.24.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomlkit==0.13.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.17.1
